refactor: restructure ChatWindow and Sidebar components; update import paths and remove unused files
